Logging out all users without refresh token
Past users don't have the refresh token in your locals storage because
this info wasn't persisted in previous version, because this when the
past users token expires the extension is not able to refresh the token.

This commit, with much shame, remove the token from the users machine
with 1.5.0 version and the refreshToken undefined.