Fix the problem on get the comment ID from event URL
I was optimistic thinking the all comments events brought the comment ID
in event url anchor, but in some cases this don't happen, with this the
extension stopped to show the event popup content when some event like that
was announced.

With this commit when an event is sign as comment but is impossible to
grab the comment ID from event URL anchor the extension assumes which the first
comment is the announced comment.