New popup and background management
This commit introduces a whole new front-end code to the event's popup
and the background.

This new front-end uses more standalone modules to small tasks instead of
rely fully on the Backbone's models, which complicated so much the
understanding of what was going on. These new modules are more simple
and easy to follow the flow.

This new front-end also includes the React on the presentation layer and
Fluxo on the action and app state management.

React is much more smarter on the rendering than Backbone view.

Fluxo brings the idea of unidirectional flow to the app, and it helps a lot
to understand what happen when some action is called, like star an event,
for example. You can follow the flow on just one place, and the presentation
layer never change the "model" state, just call actions to action handlers do
what it needs to do.

With these changes Backbone.js gonna be fully removed soon, the model
Fluxo + React + smaller modules doing smaller tasks looks much more
promising.

Sadly many interdependents changes happened and I can't split this
giant commit on smaller ones.