Centralizing the local storage access
There's no need to maintains two object to access the local storage, so
everything was centralized on the array local storage.

This commit also changes the way to star an event and mark as unread, avoiding touch the last events cache and using standalone arrays with
IDs, which can be less error-prone.