Loading the account user ID
The most reliable way to check if an event isn't created by the current
user is comparing the current user ID with the event's creator ID, but
we have a problem.

Basecamp send on the authentication endpoint an user ID that can't be
used to compare the creator, once the Basecamp give a different user ID
by account (in fact this ID isn't the user ID, it's a "account access" ID),
so we need fetch this account access ID when user click to listen a new
account on the config page.