Simplifying the build and ES6 with Babel.js
This commit removes the Grunt.js in favor to a simple shell scripts.

I believe that our build process can be something very simple.

This commit also remove the Require.js in favor of the new ES6 module, as
Babel.js doesn't implement any module loading logic, we need choose one module loader
to do this task, and the choose one was the browserfiy, that are more simple
than Require.js.